memory usege plus split exec/compiler
